+++ Api_c_sharp/AutoPulse.Shared/DTO/NotificationUpdateDTO.cs
@@ -9,6 +9,6 @@ public class NotificationUpdateDTO
     public string Type { get; set; } = "info";
     public string? UrlNavigation { get; set; }
     public int? IdAnnonce { get; set; }
-    public decimal? AncienPrix { get; set; }
-    public decimal? NouveauPrix { get; set; }
+    public double? AncienPrix { get; set; }
+    public double? NouveauPrix { get; set; }
 }
